🐛 エディタが起動しないcommitでテンプレートを挿入しない
scripts/commit-template.jsを修正。
テキストエディタでの編集が行えないcommitで、テンプレートを挿入しないように変更。

+++ scripts/commit-template.js
@@ -11,6 +11,7 @@ const TEMPLATE_FILE_LIST = [
 const [ COMMIT_MSG_FILE, COMMIT_SOURCE, SHA1 ] = (process.env.HUSKY_GIT_PARAMS || process.env.GIT_PARAMS || '').split(' ');
 const SCRIPT_PATH = './' + path.relative(process.cwd(), process.argv[1]);
 const OUTPUT_PREFIX = `${SCRIPT_PATH} > `;
+const DEFAULT_COMMENT_REGEXP = /(?=# Please enter the commit message)(?:#[^\r\n]*(?:\r\n?|\n))+$/;
 
 const getTemplate = async () => {
   const filelist = [];
@@ -42,15 +43,17 @@ const getTemplate = async () => {
     const existing = fs.readFileSync(msgFilePath).toString();
     const templateText = await getTemplate();
 
-    if (existing.substr(0, templateText.length) === templateText && /^[\r\n]?$/.test(existing.charAt(templateText.length))) {
-      console.log(`${OUTPUT_PREFIX}Commit template was duplicated`);
-    } else {
-      try {
-        fs.writeFileSync(msgFilePath, `${templateText}\n\n${existing}`);
-        console.log(`${OUTPUT_PREFIX}Commit template inserted`);
-      } catch(error) {
-        console.error(`${OUTPUT_PREFIX}[!] COMMIT_MSG_FILE can't write`);
-        console.log(`${OUTPUT_PREFIX}\n${await getTemplate()}`);
+    if (DEFAULT_COMMENT_REGEXP.test(existing)) {
+      if (existing.substr(0, templateText.length) === templateText && /^[\r\n]?$/.test(existing.charAt(templateText.length))) {
+        console.log(`${OUTPUT_PREFIX}Commit template was duplicated`);
+      } else {
+        try {
+          fs.writeFileSync(msgFilePath, `${templateText}\n\n${existing}`);
+          console.log(`${OUTPUT_PREFIX}Commit template inserted`);
+        } catch(error) {
+          console.error(`${OUTPUT_PREFIX}[!] COMMIT_MSG_FILE can't write`);
+          console.log(`${OUTPUT_PREFIX}\n${await getTemplate()}`);
+        }
       }
     }
   } catch(error) {
